Write functions to get and filter matches from database. Use DB instead of scraper on the main route. Cleanup and move some functions
